# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/endpoints/{id}" : {
      "get" : {
        "tags" : [ "monitored-endpoint-controller" ],
        "operationId" : "getEndpoint",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEndpointDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "monitored-endpoint-controller" ],
        "operationId" : "updateEndpoint",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EndpointUpdateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEndpointDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "monitored-endpoint-controller" ],
        "operationId" : "deleteEndpoint",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/endpoints" : {
      "get" : {
        "tags" : [ "monitored-endpoint-controller" ],
        "operationId" : "getEndpoints",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "environment",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "method",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "checkType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "isActive",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          }
        }, {
          "name" : "sortDir",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseDtoEndpointDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "monitored-endpoint-controller" ],
        "operationId" : "createEndpoint",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EndpointCreateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEndpointDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents" : {
      "get" : {
        "tags" : [ "incident-controller" ],
        "operationId" : "getIncidents",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "severity",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "endpointId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "startedAt"
          }
        }, {
          "name" : "sortDir",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseDtoIncidentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{id}" : {
      "get" : {
        "tags" : [ "incident-controller" ],
        "operationId" : "getIncidentById",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseIncidentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "EndpointUpdateCommand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string",
            "pattern" : "^(http|https)://.*$"
          },
          "method" : {
            "type" : "string",
            "enum" : [ "GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS" ]
          },
          "environment" : {
            "type" : "string"
          },
          "checkType" : {
            "type" : "string",
            "enum" : [ "HTTP", "TCP", "DATABASE" ]
          },
          "isActive" : {
            "type" : "boolean"
          },
          "policyId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "alertRuleIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "headers" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "requestBody" : {
            "type" : "string"
          }
        },
        "required" : [ "checkType", "id", "method", "name", "policyId", "url" ]
      },
      "ApiResponseEndpointDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/EndpointDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "EndpointDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "method" : {
            "type" : "string",
            "enum" : [ "GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS" ]
          },
          "environment" : {
            "type" : "string"
          },
          "checkType" : {
            "type" : "string",
            "enum" : [ "HTTP", "TCP", "DATABASE" ]
          },
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expectedStatusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "headers" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "requestBody" : {
            "type" : "string"
          },
          "intervalSeconds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "timeoutMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "failureThreshold" : {
            "type" : "integer",
            "format" : "int32"
          },
          "latencyThresholdMillis" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "EndpointCreateCommand" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string",
            "pattern" : "^(http|https)://.*$"
          },
          "method" : {
            "type" : "string",
            "enum" : [ "GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS" ]
          },
          "environment" : {
            "type" : "string"
          },
          "checkType" : {
            "type" : "string",
            "enum" : [ "HTTP", "TCP", "DATABASE" ]
          },
          "policyId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "alertRuleIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "headers" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "requestBody" : {
            "type" : "string"
          }
        },
        "required" : [ "checkType", "method", "name", "policyId", "url" ]
      },
...
      "ApiResponsePagedResponseDtoIncidentDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseDtoIncidentDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "IncidentDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endpointId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endpointName" : {
            "type" : "string"
          },
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "failureCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "severity" : {
            "type" : "string"
          },
          "rootCause" : {
            "type" : "string"
          },
          "failingResultIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
      "PagedResponseDtoIncidentDto" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IncidentDto"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ApiResponseIncidentDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/IncidentDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponsePagedResponseDtoEndpointDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseDtoEndpointDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PagedResponseDtoEndpointDto" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EndpointDto"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    }
  }
}